+++ great_expectations/__init__.py
@@ -1,5 +1,5 @@
 from .util import *
-from great_expectations import dataset
+from great_expectations import data_asset
 from great_expectations.data_context import get_data_context
 from .file_expectations import *
