+++ great_expectations/data_asset/__init__.py
@@ -1,3 +1,4 @@
-from .base import Dataset
+from .base import DataAsset
+from .dataset import Dataset
 from .pandas_dataset import MetaPandasDataset, PandasDataset
 from .sqlalchemy_dataset import MetaSqlAlchemyDataset, SqlAlchemyDataset
