+++ tests/test_fixtures/custom_dataset.py
@@ -1,4 +1,4 @@
-from great_expectations.dataset.pandas_dataset import PandasDataset
+from great_expectations.data_asset.pandas_dataset import PandasDataset
 
 
 class CustomPandasDataset(PandasDataset):
