+++ tests/test_parameter_substitution.py
@@ -6,20 +6,20 @@ at evaluation time, and store parameters in expectations_config
 import pytest
 import numpy as np
 
-from great_expectations.dataset import Dataset
+from great_expectations.data_asset import DataAsset
 
 
 @pytest.fixture
-def dataset():
-    return Dataset()
+def data_asset():
+    return DataAsset()
 
 
 @pytest.fixture
-def single_expectation_custom_dataset():
+def single_expectation_custom_data_asset():
 
-    class SlimCustomDataset(Dataset):
+    class SlimCustomDataAsset(DataAsset):
 
-        @Dataset.expectation('expectation_argument')
+        @DataAsset.expectation('expectation_argument')
         def expect_nothing(self, expectation_argument):
             return {
                 'success': True,
@@ -30,32 +30,32 @@ def single_expectation_custom_dataset():
                 }
             }
 
-    return SlimCustomDataset()
+    return SlimCustomDataAsset()
 
 
-def test_store_evaluation_parameter(dataset):
-    dataset.set_evaluation_parameter("my_parameter", "value")
-    assert dataset.get_evaluation_parameter("my_parameter") == "value"
+def test_store_evaluation_parameter(data_asset):
+    data_asset.set_evaluation_parameter("my_parameter", "value")
+    assert data_asset.get_evaluation_parameter("my_parameter") == "value"
 
-    dataset.set_evaluation_parameter(
+    data_asset.set_evaluation_parameter(
         "my_second_parameter", [1, 2, "value", None, np.nan])
-    assert dataset.get_evaluation_parameter("my_second_parameter") == [
+    assert data_asset.get_evaluation_parameter("my_second_parameter") == [
         1, 2, "value", None, np.nan]
 
     with pytest.raises(TypeError):
-        dataset.set_evaluation_parameter(
+        data_asset.set_evaluation_parameter(
             ["a", "list", "cannot", "be", "a", "parameter"], "value")
 
 
-def test_parameter_substitution(single_expectation_custom_dataset):
+def test_parameter_substitution(single_expectation_custom_data_asset):
     # Set our evaluation parameter from upstream
-    single_expectation_custom_dataset.set_evaluation_parameter(
+    single_expectation_custom_data_asset.set_evaluation_parameter(
         "upstream_dag_key", "upstream_dag_value")
 
     # Establish our expectation using that parameter
-    result = single_expectation_custom_dataset.expect_nothing(
+    result = single_expectation_custom_data_asset.expect_nothing(
         expectation_argument={"$PARAMETER": "upstream_dag_key"})
-    config = single_expectation_custom_dataset.get_expectations_config()
+    config = single_expectation_custom_data_asset.get_expectations_config()
 
     # Ensure our value has been substituted during evaluation, and set properly in the config
     assert result["result"]["details"]["expectation_argument"] == "upstream_dag_value"
@@ -65,13 +65,13 @@ def test_parameter_substitution(single_expectation_custom_dataset):
         "expectation_argument": {"$PARAMETER": "upstream_dag_key"}}
 
 
-def test_exploratory_parameter_substitution(single_expectation_custom_dataset):
+def test_exploratory_parameter_substitution(single_expectation_custom_data_asset):
     # Establish our expectation using a parameter provided at runtime
 
-    result = single_expectation_custom_dataset.expect_nothing(
+    result = single_expectation_custom_data_asset.expect_nothing(
         expectation_argument={"$PARAMETER": "upstream_dag_key",
                               "$PARAMETER.upstream_dag_key": "temporary_value"})
-    config = single_expectation_custom_dataset.get_expectations_config()
+    config = single_expectation_custom_data_asset.get_expectations_config()
     # Ensure our value has been substituted during evaluation, and NOT stored in the config
     assert result["result"]["details"]["expectation_argument"] == "temporary_value"
     assert "evaluation_parameters" not in config or config["evaluation_parameters"] == {
@@ -81,35 +81,35 @@ def test_exploratory_parameter_substitution(single_expectation_custom_dataset):
 
     # Evaluating the expectation without the parameter should now fail, because no parameters were set
     with pytest.raises(KeyError) as excinfo:
-        single_expectation_custom_dataset.validate(catch_exceptions=False)
+        single_expectation_custom_data_asset.validate(catch_exceptions=False)
         assert str(
             excinfo.message) == "No value found for $PARAMETER upstream_dag_key"
 
     # Setting a parameter value should allow it to succeed
-    single_expectation_custom_dataset.set_evaluation_parameter(
+    single_expectation_custom_data_asset.set_evaluation_parameter(
         "upstream_dag_key", "upstream_dag_value")
-    validation_result = single_expectation_custom_dataset.validate()
+    validation_result = single_expectation_custom_data_asset.validate()
     assert validation_result["results"][0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
 
 
-def test_validation_substitution(single_expectation_custom_dataset):
+def test_validation_substitution(single_expectation_custom_data_asset):
     # Set up an expectation using a parameter, providing a default value.
-    result = single_expectation_custom_dataset.expect_nothing(
+    result = single_expectation_custom_data_asset.expect_nothing(
         expectation_argument={"$PARAMETER": "upstream_dag_key",
                               "$PARAMETER.upstream_dag_key": "temporary_value"})
     assert result["result"]["details"]["expectation_argument"] == "temporary_value"
 
     # Provide a run-time evaluation parameter
-    validation_result = single_expectation_custom_dataset.validate(
+    validation_result = single_expectation_custom_data_asset.validate(
         evaluation_parameters={"upstream_dag_key": "upstream_dag_value"})
     assert validation_result["results"][0]["result"]["details"]["expectation_argument"] == "upstream_dag_value"
 
 
-def test_validation_parameters_returned(single_expectation_custom_dataset):
-    single_expectation_custom_dataset.expect_nothing(
+def test_validation_parameters_returned(single_expectation_custom_data_asset):
+    single_expectation_custom_data_asset.expect_nothing(
         expectation_argument={"$PARAMETER": "upstream_dag_key",
                               "$PARAMETER.upstream_dag_key": "temporary_value"})
-    validation_result = single_expectation_custom_dataset.validate(
+    validation_result = single_expectation_custom_data_asset.validate(
         evaluation_parameters={"upstream_dag_key": "upstream_dag_value"})
     assert validation_result["evaluation_parameters"] == {
         "upstream_dag_key": "upstream_dag_value"}
